added conclusie component
